requirements.txt
# Web
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
python-dotenv>=1.0,<2

# LangChain + integrations (pinned to 0.3 series)
langchain==0.3.*
langchain-core==0.3.*
langchain-text-splitters==0.3.*
langchain-community==0.3.*
langchain-upstage>=0.1.6

# Vector Store
chromadb>=0.5.5

# Optional: ensure Pydantic v2 (FastAPI dependency, included transitively)
pydantic>=2.7,<3

# HTML parsing
beautifulsoup4>=4.12,<5

# CLI progress bar
tqdm>=4.66,<5

# Web Search
tavily-python>=0.3.0

# PDF processing (for PDFMinerPDFasHTMLLoader)
pdfminer.six>=20221105

# Encoding (for tiktoken in pdf_processor.py)
tiktoken>=0.5.0

# HTTP requests (for web_search_fallback.py)
requests>=2.31,<3

# D001 전용
prometheus-client>=0.20,<1
slowapi>=0.1.4,<0.2 

